restructure module, add metadata, fill out setup.py (#2)
